Added object mapper to RouteComponentWrapper
