+++ src/App.tsx
@@ -23,7 +23,12 @@ import {
   classes,
   packageInfo
 } from './app/constants';
-import { AppLoading } from './app/AppLoading';
+import { 
+  AppLoading,
+  Offline, 
+} from './app/widgets';
+import { ReactoryRouter } from './app/router';
+
 import queryString from './components/utility/query-string';
 import './App.css';
 import { ReactoryHeader as Header } from '@reactory/client-core/components/shared/header';
@@ -83,9 +88,6 @@ const StyledRouter = styled(Router)(({ theme }: { theme: Theme }) => {
   };
 });
 
-
-
-
 if (localStorage) {
   localStorage.setItem('REACT_APP_CLIENT_KEY', REACT_APP_CLIENT_KEY);
   localStorage.setItem('REACT_APP_CLIENT_PASSWORD', REACT_APP_CLIENT_PASSWORD);
@@ -132,533 +134,6 @@ interface ReactoryRouterProps {
   header: React.ReactElement
   footer: React.ReactElement
 }
-
-
-/**
- * Wrapper component that renders inside each route to access params
- * and process componentProps templates
- */
-const RouteComponentWrapper = ({ routeDef, reactory, componentArgs, children, onComponentLoad, hasHeader = false, headerHeight = 48 }) => {
-  const params = useParams();
-  const location = useLocation();
-
-  // Process componentProps templates with actual route params
-  let processedArgs = { ...componentArgs };
-
-  // Add route params directly to component args
-  Object.keys(params).forEach(paramKey => {
-    processedArgs[paramKey] = params[paramKey];
-  });
-
-  if (routeDef.componentProps) {
-    processedArgs = { ...routeDef.componentProps };
-
-    // Process template strings in componentProps
-    Object.keys(processedArgs).forEach(key => {
-      const value = processedArgs[key];
-      if (typeof value === 'string' && value.includes('${')) {
-        try {
-          if (value.includes('::')) {
-            const [_value, transform] = value.split('::');
-            processedArgs[key] = reactory.utils.template(_value)({ route: params, location });
-            if (transform) {
-              switch (transform) {
-                case 'toInt':
-                  processedArgs[key] = parseInt(processedArgs[key]);
-                  break;
-                case 'toString':
-                  processedArgs[key] = String(processedArgs[key]);
-                  break;
-                case 'toDate':
-                  processedArgs[key] = new Date(processedArgs[key]);
-                  break;
-                case 'toBoolean':
-                  processedArgs[key] = Boolean(processedArgs[key]);
-                  break;
-                default:
-                  processedArgs[key] = processedArgs[key];
-              }
-            }
-          } else {
-            // Replace ${route.paramName} with actual param values
-            processedArgs[key] = reactory.utils.template(value)({ route: params, location });
-          }
-        } catch (error) {
-          reactory.warning(`Error processing template ${value}:`, error);
-          processedArgs[key] = value; // fallback to original value
-        }
-      }
-    });
-  }
-
-
-
-  const ReactoryComponent = reactory.getComponent(routeDef.componentFqn);
-  const NotFound = reactory.getComponent("core.NotFound");
-
-  // Calculate margin-top based on header presence and height
-  const wrapperStyle = hasHeader ? { marginTop: `${headerHeight}px` } : {};
-
-  // Add additional safety checks for component loading
-  if (!ReactoryComponent) {
-    return (
-      <div key={`route-loading-${routeDef.id}`}>
-        {children}
-        <div style={{ display: 'flex', justifyContent: 'center', padding: '10% 2rem', marginTop: `${headerHeight}px` }}>
-          <AppLoading message="Loading component..." />
-        </div>
-      </div>
-    );
-  }
-
-  try {
-    if (ReactoryComponent) {
-
-      const componentProps = {
-        reactory,
-        ...processedArgs,
-        style: {
-          marginTop: `${headerHeight}px`,
-          ...processedArgs?.style,
-        }
-      }
-
-      return (
-        <div key={`route-${routeDef.id}`}>
-          {children}
-          <ReactoryComponent {...componentProps} key={routeDef.id} />
-        </div>
-      );
-    } else {
-      // Component not found - show loading state with retry mechanism
-      return (
-        <div key={`route-not-found-${routeDef.id}`}>
-          {children}
-          <NotFound
-            key={routeDef.id}
-            message={`Component ${routeDef.componentFqn} not found for route ${routeDef.path}`}
-            waitingFor={routeDef.componentFqn}
-            args={processedArgs}
-            wait={500}
-            onFound={onComponentLoad}
-            style={{ marginTop: `${headerHeight}px` }}
-          />
-        </div>
-      );
-    }
-  } catch (routeError) {
-    reactory.error(`Error rendering component ${routeDef.componentFqn} for route ${routeDef.path}`, {
-      error: routeError,
-      routeDef,
-      params,
-      location
-    });
-    return (
-      <div key={`route-error-${routeDef.id}`}>
-        {children}
-        <NotFound
-          key={routeDef.id}
-          message={`Error rendering component ${routeDef.componentFqn} for route ${routeDef.path}`}
-          waitingFor={routeDef.componentFqn}
-          args={processedArgs}
-          wait={500}
-          onFound={onComponentLoad}
-          style={{ marginTop: `${headerHeight}px` }}
-        />
-      </div>
-    );
-  }
-};
-
-const ReactoryRouter = (props: ReactoryRouterProps) => {
-
-  const navigation = useNavigate();
-  const location = useLocation();
-  const reactory = useReactory();
-  // Remove useParams from here - it won't work at router level
-  const {
-    utils,
-    debug,
-    error,
-    warning,
-  } = reactory;
-
-  const {
-    objectMapper,
-  } = utils;
-  const { auth_validated, user, authenticating = false } = props;
-  const [routes, setRoutes] = React.useState<Reactory.Routing.IReactoryRoute[]>([]);
-  const [v, setVersion] = React.useState<number>(0);
-  const onLogin = () => {
-    // Add a small delay to ensure the main app state has been updated
-    setTimeout(() => {
-      configureRouting();
-      setVersion(v + 1);
-    }, 100);
-  };
-
-
-  const onLogout = () => {
-    // Add a small delay to ensure the main app state has been updated
-    setTimeout(() => {
-      setVersion(v + 1);
-    }, 100);
-  };
-
-  reactory.on(ReactoryApiEventNames.onLogout, onLogout);
-  reactory.on(ReactoryApiEventNames.onLogin, onLogin);
-
-  reactory.navigation = navigation;
-  reactory.location = location;
-
-
-  const configureRouting = () => {
-    debug('Configuring Routing', { auth_validated, user });
-
-    // Add safety checks to prevent route configuration during transitions
-    if (authenticating) {
-      debug('Skipping route configuration - authenticating', {
-        authenticating
-      });
-      return;
-    }
-
-    const $routes = [...reactory.getRoutes()];
-    if (reactory.utils.hashCode(JSON.stringify($routes)) !== reactory.utils.hashCode(JSON.stringify(routes))) {
-      setRoutes($routes);
-      setVersion(v + 1);
-    }
-  }
-
-  useEffect(() => {
-    configureRouting();
-  }, []);
-
-  useEffect(() => {
-    configureRouting();
-  }, [auth_validated, authenticating])
-
-  if (auth_validated === false || authenticating === true) {
-    return (<span>Validating Authentication</span>)
-  }
-
-  const ChildRoutes = [];
-  routes.forEach((routeDef: Reactory.Routing.IReactoryRoute, index: number) => {
-    reactory.log(`Configuring Route ${routeDef.path}`, { routeDef, props: props });
-    if (routeDef.redirect) {
-      // Use direct browser navigation for immediate redirect
-      debug('Executing immediate redirect:', routeDef.redirect);
-      window.location.href = routeDef.redirect;
-      return; // Skip creating a route element for redirects
-    }
-
-    let componentArgs = {};
-
-    /**
-     * If the route has args, we add them to the component args
-     * this is the secondary method of setting the props, provides
-     * additional meta data about the props to allow for potential 
-     * future features.
-     */
-    if (isArray(routeDef.args)) {
-      routeDef.args.forEach((arg) => {
-        componentArgs[arg.key] = arg.value[arg.key];
-      })
-    }
-
-    let children = [];
-    let hasHeader = false;
-    let headerHeight = 48; // Default header height
-
-    // Determine header configuration
-    if (routeDef.header) {
-      const {
-        componentFqn,
-        show = true,
-        propsMap
-      } = routeDef.header;
-
-      const headerProps: { height?: number } = routeDef.header.props || {
-        height: 48
-      };
-
-      if (show === true) {
-        const Header = reactory.getComponent<any>(componentFqn);
-        if (Header) {
-          let $props = { ...headerProps };
-          if (propsMap) {
-            $props = { ...headerProps, }
-          }
-          children.push(<Header {...$props} />);
-          hasHeader = true;
-          headerHeight = headerProps?.height || 48;
-        } else {
-          // Fallback to default header if custom header component not found
-          if (props.header) {
-            children.push(props.header);
-            hasHeader = true;
-            headerHeight = 48; // Default header height
-          }
-        }
-      }
-    } else if (props.header) {
-      children.push(props.header);
-      hasHeader = true;
-      headerHeight = 48; // Default header height
-    }
-
-    if (routeDef.public === true) {
-      // public access we don't have to check roles or auth
-      ChildRoutes.push(
-        <Route
-          key={index}
-          path={routeDef.path}
-          element={
-            <RouteComponentWrapper
-              routeDef={routeDef}
-              reactory={reactory}
-              componentArgs={componentArgs}
-              children={children}
-              onComponentLoad={() => setVersion(v + 1)}
-              hasHeader={hasHeader}
-              headerHeight={headerHeight}
-            />
-          }
-        />
-      );
-    } else {
-      // Check if user is anonymous and route is not public      
-      if (reactory.isAnon() && !routeDef.public && routeDef.path !== "/login") {
-        // If user is anonymous but route requires authentication, redirect to login
-        debug('Redirecting anonymous user to login for protected route:', routeDef.path);
-        window.location.href = "/login";
-        return; // Skip creating route element for redirected routes
-      } else {
-        debug('User is authenticated or route is public, proceeding with route configuration.', { routeDef
-        });
-      }
-
-      const hasRolesForRoute = reactory.hasRole(routeDef.roles, reactory.getUser().loggedIn.roles) === true;
-
-      if (!reactory.isAnon() && hasRolesForRoute === false) {
-        const NotFoundComponent = reactory.getComponent("core.NotFound");
-        if (NotFoundComponent) {
-          children.push(React.createElement(NotFoundComponent as any, {
-            key: "not-found-permissions",
-            message: "You don't have sufficient permissions to access this route.",
-            link: routeDef.path,
-            wait: 500
-          }));
-        }
-        ChildRoutes.push(
-          <Route
-            key={index}
-            path={routeDef.path}
-            element={<React.Fragment>{children}</React.Fragment>}
-          />
-        );
-      } else {
-        if (auth_validated === true && hasRolesForRoute === true) {
-          ChildRoutes.push(
-            <Route
-              key={index}
-              path={routeDef.path}
-              element={
-                <RouteComponentWrapper
-                  routeDef={routeDef}
-                  reactory={reactory}
-                  componentArgs={componentArgs}
-                  children={children}
-                  onComponentLoad={() => setVersion(v + 1)}
-                  hasHeader={hasHeader}
-                  headerHeight={headerHeight}
-                />
-              }
-            />
-          );
-        } else {
-          debugger;
-          const hasRefreshed: boolean = localStorage.getItem('hasRefreshed') === 'true';
-
-          //auth token not validated yet in process of checking
-          //@ts-ignore
-          if (auth_validated === false || authenticating === true) {
-            children.push(<Typography style={{ display: 'flex', justifyContent: 'center', padding: '10% 2rem' }} variant='h5'>Please wait while we validate your access token...</Typography>);
-          } else {
-            // Handle login redirect - if user is anonymous and has refreshed, redirect to login
-            if (hasRefreshed && reactory.isAnon() && routeDef.path !== "/login") {
-              debug('Redirecting to login after refresh');
-              localStorage.removeItem('hasRefreshed');
-              window.location.href = "/login";
-              return; // Skip creating route element for redirected routes
-            }
-          }
-
-          ChildRoutes.push(
-            <Route
-              key={index}
-              path={routeDef.path}
-              element={<React.Fragment>{children}</React.Fragment>}
-            />
-          );
-        }
-      }
-    }
-
-    if (props.footer && routeDef.footer) {
-      const {
-        componentFqn,
-        show = true,
-        props = {},
-        propsMap
-      } = routeDef.footer;
-
-      if (show === true) {
-        const Footer = reactory.getComponent<any>(componentFqn);
-        if (Footer) {
-          let $props = { ...props };
-          if (propsMap) {
-            $props = { ...props, }
-          }
-          // Footer should be added to children in RouteComponentWrapper if needed
-        } else {
-          // use the default footer
-          // Footer should be added to children in RouteComponentWrapper if needed
-        }
-      }
-    }
-  });
-
-  const NotFoundElement = reactory.getComponent<any>("core.NotFound");
-  ChildRoutes.push(<Route key={ChildRoutes.length + 1} path="*" element={<NotFoundElement />} />)
-  return (
-    <Routes key={'reactory-router-routes-' + v}>
-      {ChildRoutes}
-    </Routes>
-  )
-}
-
-
-
-const Offline = (props: { onOfflineChanged: (isOffline: boolean) => void }) => {
-
-  const TM_BASE_DEFAULT: number = 45000;
-  const reactory = useReactory();
-  const { onOfflineChanged } = props;
-  const [timeout_base, setTimeoutBase] = React.useState<number>(TM_BASE_DEFAULT);
-  const [offline, setOfflineStatus] = React.useState<boolean>(false);
-
-  let timeoutMS: number = 45000;
-  let totals = { error: 0, slow: 0, ok: 0, total: 0 };
-  let last_slow = null;
-
-
-  const getApiStatus = async (): Promise<void> => {
-    const started = Date.now();
-
-    try {
-      const apiStatus = await reactory.status({ emitLogin: false, forceLogout: false });
-      const done = Date.now();
-      const api_ok = apiStatus.status === 'API OK'
-      setOfflineStatus(!api_ok);
-      if (offline !== !api_ok) {
-        onOfflineChanged(!api_ok);
-      }
-
-      let isSlow = false;
-
-      const newLast = {
-        when: started,
-        pingMS: done - started,
-      };
-
-      timeoutMS = timeout_base;
-
-      //if our ping timeout is slow
-      if (newLast.pingMS > 4000 && totals.total > 10) {
-        last_slow = done;
-        timeoutMS = timeout_base * 1.25;
-      }
-
-      //if our ping time is really low
-      if (newLast.pingMS > 7000 && totals.total > 10) {
-        isSlow = true;
-        timeoutMS = timeout_base * 1.5;
-      }
-
-      let next_tm_base = TM_BASE_DEFAULT;
-      if (totals.total > 5) {
-        let avg: number = (totals.ok * 100) / totals.total;
-        if (avg > 90) next_tm_base = TM_BASE_DEFAULT * 1.30
-        if (avg > 95) next_tm_base = TM_BASE_DEFAULT * 1.5;
-        if (avg > 98) next_tm_base = TM_BASE_DEFAULT * 2.5;
-      }
-
-      const newTotals = {
-        error: totals.error,
-        slow: isSlow ? totals.slow + 1 : totals.slow,
-        ok: api_ok ? totals.ok + 1 : totals.ok,
-        total: totals.total + 1,
-      };
-
-      totals = newTotals;
-
-      //@ts-ignore
-      reactory.stat(`user-session-api-status-totals`, { user_id: reactory.getUser().id, ...totals, ...newLast });
-      reactory.emit('onApiStatusTotalsChange', { ...totals, ...newLast, api_ok, isSlow });
-
-      if (next_tm_base !== timeout_base) setTimeoutBase(next_tm_base);
-
-      setTimeout(() => {
-        void getApiStatus();
-      }, timeoutMS);
-
-      reactory.debug(`Client Ping Totals:`, { totals: newTotals, nextIn: timeoutMS });
-    } catch (error) {
-      reactory.error(`Error while fetching api status`, { error: error });
-      setOfflineStatus(true);
-      onOfflineChanged(true);
-
-      totals = {
-        error: totals.error + 1,
-        slow: totals.slow,
-        ok: totals.ok + 1,
-        total: totals.total + 1,
-      };
-
-      setTimeout(() => {
-        void getApiStatus();
-      }, timeoutMS);
-    }
-  };
-
-
-  useEffect(() => {
-    void getApiStatus();
-  }, []);
-
-  if (offline === false) return null;
-  return (
-    <React.Fragment>
-      <Box style={{ margin: 'auto', textAlign: 'center', paddingTop: '100px' }}>
-        <Typography variant="h1" style={{ color: reactory.muiTheme.palette.error.main }}>
-          <Icon style={{ fontSize: '100px' }}>cloud_off</Icon>
-        </Typography>
-        <Typography variant="body2">
-          We are unable to connect you to our service at this time.
-          This may be due to a poor internet connection or your
-          device is currently offline.
-        </Typography>
-
-        <Typography variant="body2">
-          This message will disappear as soon as we are able to establish a connection.
-          If you accessed the system with an email link, please retry using this link in a few moments.
-        </Typography>
-      </Box>
-    </React.Fragment>
-  )
-}
-
 
 const dependencies = [
   'core.Loading@1.0.0',
@@ -974,10 +449,8 @@ export const ReactoryHOC = (props: ReactoryHOCProps) => {
         reactory.on(ReactoryApiEventNames.onLogin, onLogin)
         reactory.on(ReactoryApiEventNames.onApiStatusUpdate, onApiStatusUpdate);
         reactory.on(ReactoryApiEventNames.onRouteChanged, onRouteChanged);
-        reactory.on(ReactoryApiEventNames.onThemeChanged, onThemeChanged);
-
+        reactory.on(ReactoryApiEventNames.onThemeChanged, onThemeChanged);        
         const query = queryString.parse(window.location.search);
-
         reactory.queryObject = query;
         reactory.queryString = window.location.search;
         reactory.objectToQueryString = queryString.stringify;
