--- src/app/AppLoading/index.ts
@@ -1,1 +1,0 @@
-export { default as AppLoading } from './AppLoadingWidget';
