+++ src/app/widgets/index.ts
@@ -1,0 +1,2 @@
+export { default as AppLoading } from './AppLoading';
+export { default as Offline } from './Offline';
